Novas implementações na aula de 08/05

+++ 07_ProjetoMyRh_WebApp/ProjetoMyRh/ProjetoMyRh.AppWeb/Models/Entities/Cargo.cs
@@ -19,7 +19,7 @@ namespace ProjetoMyRh.AppWeb.Models.Entities
 
         [DisplayName("Tipo de Salário")]
         [Range(1, 2, ErrorMessage = "O tipo de Salário deve ser MENSAL ou POR HORA")]
-        public int TipoSalario { get; set; }
+        public System.Int16 TipoSalario { get; set; }
         public Area? AreaAtuacao { get; set; }
 
     }
